[Mod] Open various context menus on right mouse button release as it should be done, not on button down.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@22264 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Globals.h
@@ -43,7 +43,7 @@ protected:
 	CModControlView &m_parent;
 	HWND m_hWndView = nullptr;
 	HWND m_lastFocusItem = nullptr;
-	LONG m_nLockCount = 0;
+	int m_nLockCount = 0;
 	bool m_initialized = false;
 
 public:
